backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-costmanagement==4.0.0
google-cloud-billing==1.11.0
firebase-admin==6.4.0
APScheduler==3.10.4
